backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv

sqlalchemy
psycopg2-binary

pydantic

passlib[bcrypt]
python-jose[cryptography]

python-multipart

groq

alembic
pydantic [email
cloudinary
argon2_cffi